fix: use 32-bit arithmetic in MPQ crypt routines for LP64 platforms
On Linux x86_64, unsigned long is 64-bit while MPQ crypto relies on
32-bit wrapping arithmetic. This made DecryptData/HashString produce
garbage hash tables, so freeze failed with (keyfile) not found.
Use uint32_t throughout mpqcrypt.cpp/h.

+++ mpqprt/mpq/mpqcrypt.h
@@ -8,15 +8,15 @@
 #include "mpqtypes.h"
 #include <functional>
 
-void EncryptData(void *lpbyBuffer, unsigned long dwLength, unsigned long dwKey);
-void DecryptData(void *lpbyBuffer, unsigned long dwLength, unsigned long dwKey);
+void EncryptData(void *lpbyBuffer, uint32_t dwLength, uint32_t dwKey);
+void DecryptData(void *lpbyBuffer, uint32_t dwLength, uint32_t dwKey);
 
 #define MPQ_HASH_TABLE_OFFSET	0
 #define MPQ_HASH_NAME_A	1
 #define MPQ_HASH_NAME_B	2
 #define MPQ_HASH_FILE_KEY	3
 
-unsigned long HashString(const char *lpszString, unsigned long dwHashType);
+uint32_t HashString(const char *lpszString, uint32_t dwHashType);
 uint32_t GetFileDecryptKey(const void *buffer, uint32_t bufferSize, uint32_t expectedFirstDword,
                            const std::function<bool(const void *)> &validator);
